+++ UnitTests/BraintreeCardTests/BTCardClient_Tests.swift
@@ -5,15 +5,18 @@ import XCTest
 
 class BTCardClient_Tests: XCTestCase {
     let customerInputValidationErrorKey: String = "BTCustomerInputBraintreeValidationErrorsKey"
+    let sandboxAuthorization: String = "sandbox_9dbg82cq_dcpspy2brwdjr3qn"
+    
     // MARK: - ClientAPI
     
     func testTokenization_postsCardDataToClientAPI() {
         let expectation = self.expectation(description: "Tokenize Card")
 
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
 
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -54,10 +57,11 @@ class BTCardClient_Tests: XCTestCase {
     func testTokenization_whenAuthInsightIsNotRequested_postsCardDataWithoutAuthInsight() {
         let expectation = self.expectation(description: "Tokenize Card")
 
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -100,11 +104,12 @@ class BTCardClient_Tests: XCTestCase {
             ]
         ]
 
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         mockAPIClient.cannedResponseBody = BTJSON(value: mockTokenizeResponse)
 
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -133,11 +138,12 @@ class BTCardClient_Tests: XCTestCase {
 
         let mockError = NSError(domain: "TestErrorDomain", code: 1, userInfo: nil)
 
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         mockAPIClient.cannedResponseError = mockError
 
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -156,7 +162,7 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testTokenization_whenTokenizationEndpointReturns422_callCompletionWithValidationError() {
-        let stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
+        var stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
         stubAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         let stubJSONResponse = BTJSON(
             value: [
@@ -182,7 +188,9 @@ class BTCardClient_Tests: XCTestCase {
             BTCoreConstants.jsonResponseBodyKey: stubJSONResponse
         ])
         stubAPIClient.cannedResponseError = stubError
-        let cardClient = BTCardClient(apiClient: stubAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = stubAPIClient
+        
 
         let card = BTCard(
             number: "4111111111111111",
@@ -213,7 +221,7 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testTokenization_whenTokenizationEndpointReturns422AndCode81724_callCompletionWithValidationError() {
-        let stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
+        var stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
         stubAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         let stubJSONResponse = BTJSON(
             value: [
@@ -239,7 +247,8 @@ class BTCardClient_Tests: XCTestCase {
             BTCoreConstants.jsonResponseBodyKey: stubJSONResponse
         ])
         stubAPIClient.cannedResponseError = stubError
-        let cardClient = BTCardClient(apiClient: stubAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = stubAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -271,7 +280,7 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testTokenization_whenGraphQLTokenizationEndpointReturns422AndCode81724_callsCompletionWithValidationError() {
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -317,7 +326,8 @@ class BTCardClient_Tests: XCTestCase {
             BTCoreConstants.jsonResponseBodyKey: stubJSONResponse
         ])
         mockAPIClient.cannedResponseError = stubError
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
         
         let card = BTCard(
             number: "4111111111111111",
@@ -346,10 +356,11 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testTokenization_whenTokenizationEndpointReturnsAnyNon422Error_callCompletionWithError() {
-        let stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
+        var stubAPIClient = MockAPIClient(authorization: TestClientTokenFactory.validClientToken)!
         stubAPIClient.cannedResponseError = NSError(domain: BTHTTPError.errorDomain, code: BTHTTPError.clientError([:]).errorCode, userInfo: nil)
         stubAPIClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
-        let cardClient = BTCardClient(apiClient: stubAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = stubAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -371,8 +382,10 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testMetaParameter_whenTokenizationIsSuccessful_isPOSTedToServer() {
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
+        
         let card = BTCard(
             number: "4111111111111111",
             expirationMonth: "12",
@@ -401,10 +414,12 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testAnalyticsEvent_whenTokenizationSucceeds_isSent() {
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [:] as [String?: Any])
 
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
+        
         let card = BTCard(
             number: "4111111111111111",
             expirationMonth: "12",
@@ -423,7 +438,7 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testAnalyticsEvent_whenTokenizationFails_isSent() {
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         let stubJSONResponse = BTJSON(
             value: [
                 "error" : [
@@ -449,7 +464,8 @@ class BTCardClient_Tests: XCTestCase {
         ])
         mockAPIClient.cannedResponseError = stubError
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [:] as [String?: Any])
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -471,7 +487,7 @@ class BTCardClient_Tests: XCTestCase {
     // MARK: - GraphQL API
 
     func testTokenization_whenAuthInsightRequestedIsTrue_andMerchantAccountIDIsNil_returnsError() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -479,7 +495,8 @@ class BTCardClient_Tests: XCTestCase {
             ] as [String: Any]
         ])
         
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -503,7 +520,7 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testTokenization_whenGraphQLIsEnabled_postsCardDataToGraphQLAPI() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -511,7 +528,8 @@ class BTCardClient_Tests: XCTestCase {
             ] as [String: Any]
         ])
 
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -542,10 +560,12 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testTokenization_whenGraphQLIsDisabled_postsCardDataToGatewayAPI() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [] as [Any?])
         
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
+        
         let card = BTCard(
             number: "4111111111111111",
             expirationMonth: "12",
@@ -575,7 +595,7 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testTokenization_whenGraphQLFeatureIsNotEnabled_postsCardDataToGatewayAPI() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -583,7 +603,9 @@ class BTCardClient_Tests: XCTestCase {
             ] as [String: Any]
         ])
         
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
+        
         let card = BTCard(
             number: "4111111111111111",
             expirationMonth: "12",
@@ -604,7 +626,7 @@ class BTCardClient_Tests: XCTestCase {
     }
     
     func testTokenization_whenGraphQLIsEnabledAndTokenizationIsSuccessful_returnsACardNonce() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -637,7 +659,8 @@ class BTCardClient_Tests: XCTestCase {
             ] as [String: Any]
         )
 
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -675,7 +698,7 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testAnalyticsEvent_whenTokenizationSucceedsWithGraphQL_isSent() {
-        let mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockApiClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockApiClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -706,7 +729,8 @@ class BTCardClient_Tests: XCTestCase {
             "extensions": [] as [Any?]
         ] as [String: Any])
 
-        let cardClient = BTCardClient(apiClient: mockApiClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockApiClient
 
         let card = BTCard(
             number: "4111111111111111",
@@ -727,7 +751,7 @@ class BTCardClient_Tests: XCTestCase {
     }
 
     func testAnalyticsEvent_whenTokenizationFailsWithGraphQL_isSent() {
-        let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        var mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "graphQL": [
                 "url": "graphql://graphql",
@@ -756,7 +780,8 @@ class BTCardClient_Tests: XCTestCase {
             BTCoreConstants.jsonResponseBodyKey: stubJSONResponse
         ])
         mockAPIClient.cannedResponseError = stubError
-        let cardClient = BTCardClient(apiClient: mockAPIClient)
+        var cardClient = BTCardClient(authorization: sandboxAuthorization)
+        cardClient.apiClient = mockAPIClient
 
         let card = BTCard(
             number: "4111111111111111",